requirements.txt
# GS Api
requests

# Converter
pycryptodome
numpy

# Flatbuffer fixer
bson